requirements.txt
apscheduler
cloudscraper
flask
google-genai
openai
psycopg2-binary
python-dotenv
pytz
requests
